[WFLY-10522] Support Eclipse MicroProfile Config
Add the microprofile-config-smallrye subsystem that provides support for
the Eclipse MicroProfile Config using the SmallRye implementation at
https://github.com/smallrye/smallrye-config.

The integrated version is 1.3.0.Final that supports MicroProfile Config
1.3.

The name of the subsystem is `microprofile-config-smallrye`.
It is added by default to *all* WildFly configuration (standalone and
domain).

JIRA: https://issues.jboss.org/browse/WFLY-10522